+++ config/tests.sh
@@ -29,15 +29,17 @@ run_for_framework() {
         python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_pytorch_multiprocessing.py
         python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_training_with_no_grad_updates.py
         python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_pytorch_disengage_smdebug.py
+	python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_pytorch_get_hook.py
       elif [ "$1" = "tensorflow" ] ; then
         python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow_integration.py
         python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow_error_handling_agent.py
       elif [ "$1" = "tensorflow2" ] ; then
-        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow2_error_handling_agent.py
-        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow2_gradtape_integration.py
-        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow2_integration.py
-        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow2_bert.py
-        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  tests/zero_code_change/test_tensorflow_disengage_smdebug.py
+        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_1$1.html tests/zero_code_change/test_tensorflow2_error_handling_agent.py
+        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_2$1.html tests/zero_code_change/test_tensorflow2_gradtape_integration.py
+        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_3$1.html tests/zero_code_change/test_tensorflow2_integration.py
+        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_4$1.html tests/zero_code_change/test_tensorflow2_bert.py
+        python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_5$1.html tests/zero_code_change/test_tensorflow_disengage_smdebug.py
+	python -m pytest ${code_coverage_smdebug:+--cov=./ --cov-append}  --html=$REPORT_DIR/report_6$1.html tests/zero_code_change/test_tensorflow2_get_hook.py
 
       fi
